Fix keyid calculation issues
The spec currently still requires keyids to be calculated from the key
content: this is dumb but it is in the spec.

python-tuf/securesystemslib does not (and can't) make sure the keyid
matches the key content: do that manually in the cases where we modify
the key content.
